+++ lib/pages/transaction.dart
@@ -18,13 +18,12 @@ import 'package:waterflyiii/generated/l10n/app_localizations.dart';
 import 'package:waterflyiii/generated/swagger_fireflyiii_api/firefly_iii.swagger.dart';
 import 'package:waterflyiii/layout.dart';
 import 'package:waterflyiii/notificationlistener.dart';
-import 'package:waterflyiii/pages/transaction/attachments.dart';
 import 'package:waterflyiii/pages/transaction/delete.dart';
+import 'package:waterflyiii/pages/transaction/headersection.dart';
 import 'package:waterflyiii/pages/transaction/tags.dart';
 import 'package:waterflyiii/settings.dart';
 import 'package:waterflyiii/timezonehandler.dart';
 import 'package:waterflyiii/widgets/autocompletetext.dart';
-import 'package:waterflyiii/widgets/input_number.dart';
 import 'package:waterflyiii/widgets/materialiconbutton.dart';
 
 final Logger log = Logger("Pages.Transaction");
@@ -1238,162 +1237,6 @@ class _TransactionPageState extends State<TransactionPage>
         readOnly: _tx.reconciled && _tx.initiallyReconciled,
       ),
     );
-    /*
-    // Title
-    childs.add(
-      Row(
-        children: <Widget>[
-          TransactionTitle(
-            textController: _titleTextController,
-            focusNode: _titleFocusNode,
-          ),
-          const SizedBox(width: 12),
-          badges.Badge(
-            badgeContent: Text(
-              _attachments?.length.toString() ?? "..",
-              style: Theme.of(context).textTheme.labelMedium!.copyWith(
-                color: Theme.of(context).colorScheme.onSurfaceVariant,
-              ),
-            ),
-            showBadge: _hasAttachments,
-            badgeStyle: badges.BadgeStyle(
-              badgeColor: Theme.of(context).colorScheme.surfaceContainerHighest,
-            ),
-            badgeAnimation: const badges.BadgeAnimation.scale(
-              animationDuration: animDurationEmphasized,
-              curve: animCurveEmphasized,
-            ),
-            child: MaterialIconButton(
-              icon: Icons.attach_file,
-              tooltip: S.of(context).transactionAttachments,
-              onPressed: () async {
-                final List<AttachmentRead> dialogAttachments =
-                    _attachments ?? <AttachmentRead>[];
-                await showDialog<List<AttachmentRead>>(
-                  context: context,
-                  builder: (BuildContext context) => AttachmentDialog(
-                    attachments: dialogAttachments,
-                    transactionId: _transactionJournalIDs.firstWhereOrNull(
-                      (String? element) => element != null,
-                    ),
-                  ),
-                );
-                setState(() {
-                  _attachments = dialogAttachments;
-                  _hasAttachments = _attachments?.isNotEmpty ?? false;
-                });
-              },
-            ),
-          ),
-        ],
-      ),
-    );
-    childs.add(hDivider);
-
-    // Amount, Date & Time
-    childs.add(
-      // Date/Time select might overflow, so we need to be able to scroll horizontally.
-      SingleChildScrollView(
-        scrollDirection: .horizontal,
-        child: Row(
-          children: <Widget>[
-            SizedBox(
-              width: 130,
-              child: Center(
-                child: NumberInput(
-                  icon: _localCurrency != null
-                      ? SizedBox(
-                          width: 24,
-                          height: 32,
-                          child: FittedBox(
-                            child: Text(_localCurrency!.attributes.symbol),
-                          ),
-                        )
-                      : const Icon(Icons.monetization_on),
-                  hintText:
-                      _localCurrency?.zero() ??
-                      NumberFormat.currency(decimalDigits: 2).format(0),
-                  decimals: _localCurrency?.attributes.decimalPlaces ?? 2,
-                  //style: Theme.of(context).textTheme.headlineLarge,
-                  controller: _localAmountTextController,
-                  disabled:
-                      _savingInProgress ||
-                      _split ||
-                      (_reconciled && _initiallyReconciled),
-                  onChanged: (String string) =>
-                      _localAmounts[0] = double.tryParse(string) ?? 0,
-                ),
-              ),
-            ),
-            vDivider,
-            IntrinsicWidth(
-              child: TextFormField(
-                controller: _dateTextController,
-                decoration: const InputDecoration(
-                  //prefixIcon: Icon(Icons.calendar_month),
-                  border: OutlineInputBorder(),
-                ),
-                readOnly: true,
-                enabled: !_savingInProgress,
-                onTap: () async {
-                  final DateTime? pickedDate = await showDatePicker(
-                    context: context,
-                    initialDate: _date,
-                    locale: Locale(
-                      Intl.defaultLocale!.split("_").first,
-                      Intl.defaultLocale!.split("_").last,
-                    ),
-                    firstDate: DateTime(2000),
-                    lastDate: DateTime(2101),
-                  );
-
-                  if (pickedDate == null) {
-                    return;
-                  }
-
-                  setState(() {
-                    _date = tz.TZDateTime.from(
-                      _date.copyWith(
-                        year: pickedDate.year,
-                        month: pickedDate.month,
-                        day: pickedDate.day,
-                      ),
-                      _date.location,
-                    );
-                    _dateTextController.text = DateFormat.yMMMd().format(_date);
-                  });
-                },
-              ),
-            ),
-            vDivider,
-            IntrinsicWidth(
-              child: TextFormField(
-                controller: _timeTextController,
-                decoration: const InputDecoration(border: OutlineInputBorder()),
-                readOnly: true,
-                enabled: !_savingInProgress,
-                onTap: () async {
-                  final TimeOfDay? pickedTime = await showTimePicker(
-                    context: context,
-                    initialTime: _date.getTimeOfDay(),
-                  );
-
-                  if (pickedTime == null) {
-                    return;
-                  }
-
-                  setState(() {
-                    _date = _date.setTimeOfDay(pickedTime);
-                    _timeTextController.text = DateFormat.Hm().format(_date);
-                  });
-                },
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-    */
     childs.add(hDivider);
 
     // Source Account, Destination Account & floating type element
@@ -2290,66 +2133,6 @@ class TransactionDeleteButton extends StatelessWidget {
   }
 }
 
-class TransactionTitle extends StatelessWidget {
-  const TransactionTitle({
-    super.key,
-    required this.textController,
-    required this.focusNode,
-  });
-
-  final TextEditingController textController;
-  final FocusNode focusNode;
-
-  @override
-  Widget build(BuildContext context) {
-    final Logger log = Logger("Pages.Transaction.Title");
-
-    CancelableOperation<Response<AutocompleteTransactionArray>>? fetchOp;
-
-    log.finest(() => "build()");
-    return Expanded(
-      child: AutoCompleteText<String>(
-        disabled: _savingInProgress,
-        labelText: S.of(context).transactionFormLabelTitle,
-        labelIcon: Icons.receipt_long,
-        textController: textController,
-        focusNode: focusNode,
-        optionsBuilder: (TextEditingValue textEditingValue) async {
-          try {
-            unawaited(fetchOp?.cancel());
-
-            final FireflyIii api = context.read<FireflyService>().api;
-            fetchOp =
-                CancelableOperation<
-                  Response<AutocompleteTransactionArray>
-                >.fromFuture(
-                  api.v1AutocompleteTransactionsGet(
-                    query: textEditingValue.text,
-                  ),
-                );
-            final Response<AutocompleteTransactionArray>? response =
-                await fetchOp?.valueOrCancellation();
-            if (response == null) {
-              // Cancelled
-              return const Iterable<String>.empty();
-            }
-            apiThrowErrorIfEmpty(response, context.mounted ? context : null);
-
-            return response.body!.map((AutocompleteTransaction e) => e.name);
-          } catch (e, stackTrace) {
-            log.severe(
-              "Error while fetching autocomplete from API",
-              e,
-              stackTrace,
-            );
-            return const Iterable<String>.empty();
-          }
-        },
-      ),
-    );
-  }
-}
-
 class TransactionNote extends StatelessWidget {
   const TransactionNote({super.key, required this.textController});
 
@@ -2630,338 +2413,4 @@ class _AttachmentButtonState extends State<AttachmentButton> {
       ),
     );
   }
-}
-
-class DateTimePicker extends StatefulWidget {
-  const DateTimePicker({
-    super.key,
-    required this.initialDateTime,
-    required this.onDateTimeChanged,
-  });
-
-  final tz.TZDateTime initialDateTime;
-  final ValueChanged<tz.TZDateTime> onDateTimeChanged;
-
-  @override
-  State<DateTimePicker> createState() => _DateTimePickerState();
-}
-
-class _DateTimePickerState extends State<DateTimePicker> {
-  late tz.TZDateTime _selectedDateTime;
-  late TextEditingController _dateTextController;
-  late TextEditingController _timeTextController;
-
-  final Logger log = Logger("Pages.Transaction.DateTimePicker");
-
-  @override
-  void initState() {
-    super.initState();
-
-    log.finest(() => "initState()");
-    _selectedDateTime = widget.initialDateTime;
-    _dateTextController = TextEditingController(
-      text: DateFormat.yMMMd().format(_selectedDateTime),
-    );
-    _timeTextController = TextEditingController(
-      text: DateFormat.Hm().format(_selectedDateTime),
-    );
-  }
-
-  @override
-  void dispose() {
-    _dateTextController.dispose();
-    _timeTextController.dispose();
-
-    super.dispose();
-  }
-
-  Future<void> _pickDate() async {
-    final DateTime? pickedDate = await showDatePicker(
-      context: context,
-      initialDate: _selectedDateTime,
-      locale: Locale(
-        Intl.defaultLocale!.split('_').first,
-        Intl.defaultLocale!.split('_').last,
-      ),
-      firstDate: DateTime(2000),
-      lastDate: DateTime(2101),
-    );
-
-    if (pickedDate == null) {
-      return;
-    }
-
-    setState(() {
-      _selectedDateTime = tz.TZDateTime.from(
-        _selectedDateTime.copyWith(
-          year: pickedDate.year,
-          month: pickedDate.month,
-          day: pickedDate.day,
-        ),
-        _selectedDateTime.location,
-      );
-      _dateTextController.text = DateFormat.yMMMd().format(_selectedDateTime);
-      widget.onDateTimeChanged(_selectedDateTime);
-    });
-  }
-
-  Future<void> _pickTime() async {
-    final TimeOfDay? pickedTime = await showTimePicker(
-      context: context,
-      initialTime: _selectedDateTime.getTimeOfDay(),
-    );
-
-    if (pickedTime == null) {
-      return;
-    }
-
-    setState(() {
-      _selectedDateTime = _selectedDateTime.setTimeOfDay(pickedTime);
-      _timeTextController.text = DateFormat.Hm().format(_selectedDateTime);
-      widget.onDateTimeChanged(_selectedDateTime);
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    log.finest("build()");
-
-    return Row(
-      children: <Widget>[
-        IntrinsicWidth(
-          child: TextFormField(
-            enabled: !_savingInProgress,
-            controller: _dateTextController,
-            decoration: InputDecoration(
-              //prefixIcon: Icon(Icons.calendar_month),
-              border: const OutlineInputBorder(),
-              filled: _savingInProgress,
-            ),
-            readOnly: true,
-            onTap: _pickDate,
-          ),
-        ),
-        const SizedBox(width: 16),
-        IntrinsicWidth(
-          child: TextFormField(
-            enabled: !_savingInProgress,
-            controller: _timeTextController,
-            decoration: InputDecoration(
-              border: const OutlineInputBorder(),
-              filled: _savingInProgress,
-            ),
-            readOnly: true,
-            onTap: _pickTime,
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class TransactionHeaderSection extends StatelessWidget {
-  final TransactionState tx;
-  final TextEditingController totalAmountTC;
-  final FocusNode totalAmountFN;
-  final bool saving;
-  final bool readOnly;
-
-  const TransactionHeaderSection({
-    super.key,
-    required this.tx,
-    required this.totalAmountTC,
-    required this.totalAmountFN,
-    required this.saving,
-    required this.readOnly,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final Logger log = Logger("Pages.Transaction.TransactionHeaderSection");
-
-    log.finest("build()");
-    final bool isSplit = tx.split;
-
-    return Column(
-      children: <Widget>[
-        Row(
-          children: <Widget>[
-            _buildTitle(context, isSplit),
-            const SizedBox(width: 12),
-            _buildAttachmentButton(context),
-          ],
-        ),
-        const SizedBox(height: 16),
-        SingleChildScrollView(
-          scrollDirection: .horizontal,
-          child: Row(
-            children: <Widget>[
-              _buildAmount(context, isSplit),
-              const SizedBox(width: 16),
-              _buildDate(context),
-              const SizedBox(width: 16),
-              _buildTime(context),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildTitle(BuildContext context, bool isSplit) {
-    if (isSplit) {
-      return TransactionTitle(
-        textController: tx.groupTitleTC,
-        focusNode: tx.groupTitleFN,
-      );
-    } else {
-      final TransactionSplitState split = tx.splits.first;
-
-      return TransactionTitle(
-        textController: split.titleTC,
-        focusNode: split.titleFN,
-      );
-    }
-  }
-
-  Widget _buildAttachmentButton(BuildContext context) {
-    return badges.Badge(
-      badgeContent: Text(
-        tx.attachments == null ? ".." : tx.attachments!.length.toString(),
-        style: Theme.of(context).textTheme.labelMedium!.copyWith(
-          color: Theme.of(context).colorScheme.onSurfaceVariant,
-        ),
-      ),
-      showBadge: tx.hasAttachments,
-      badgeStyle: badges.BadgeStyle(
-        badgeColor: Theme.of(context).colorScheme.surfaceContainerHighest,
-      ),
-      badgeAnimation: const badges.BadgeAnimation.scale(
-        animationDuration: animDurationEmphasized,
-        curve: animCurveEmphasized,
-      ),
-      child: MaterialIconButton(
-        icon: Icons.attach_file,
-        tooltip: S.of(context).transactionAttachments,
-        onPressed: () async {
-          final List<AttachmentRead> dialogAttachments =
-              tx.attachments ?? <AttachmentRead>[];
-          await showDialog<List<AttachmentRead>>(
-            context: context,
-            builder: (BuildContext context) => AttachmentDialog(
-              attachments: dialogAttachments,
-              transactionId: tx.splits.first.journalID,
-            ),
-          );
-          tx.attachments = dialogAttachments;
-        },
-      ),
-    );
-  }
-
-  Widget _buildAmount(BuildContext context, bool isSplit) {
-    if (tx.splits.isEmpty) {
-      return const SizedBox.shrink();
-    }
-
-    late final Widget input;
-
-    if (!isSplit) {
-      // SINGLE TX --> editable
-      final TransactionSplitState split = tx.splits.first;
-
-      input = _buildNumberInput(
-        controller: split.localAmountTC,
-        enabled: !saving && !readOnly,
-        onChanged: (String val) {
-          log.finest(() => "_buildAmount: onChanged($val)");
-          split.localAmount = double.tryParse(val) ?? 0;
-        },
-        currency: tx.localCurrency,
-      );
-    } else {
-      // SPLIT TX --> derived total (readonly)
-      input = _buildNumberInput(
-        controller: totalAmountTC,
-        enabled: false,
-        currency: tx.localCurrency,
-      );
-    }
-    return SizedBox(width: 135, child: input);
-  }
-
-  Widget _buildNumberInput({
-    required TextEditingController controller,
-    required bool enabled,
-    void Function(String)? onChanged,
-    required CurrencyRead currency,
-  }) {
-    return NumberInput(
-      icon: SizedBox(
-        width: 24,
-        height: 32,
-        child: FittedBox(child: Text(currency.attributes.symbol)),
-      ),
-      hintText: currency.zero(),
-      decimals: currency.attributes.decimalPlaces ?? 2,
-      controller: controller,
-      disabled: !enabled,
-      onChanged: onChanged,
-    );
-  }
-
-  Widget _buildDate(BuildContext context) => IntrinsicWidth(
-    child: TextFormField(
-      controller: tx.dateTC,
-      decoration: const InputDecoration(border: OutlineInputBorder()),
-      readOnly: true,
-      enabled: !saving,
-      onTap: () async {
-        final DateTime? pickedDate = await showDatePicker(
-          context: context,
-          initialDate: tx.date,
-          locale: Locale(
-            Intl.defaultLocale!.split("_").first,
-            Intl.defaultLocale!.split("_").last,
-          ),
-          firstDate: DateTime(2000),
-          lastDate: DateTime(2101),
-        );
-
-        if (pickedDate == null) {
-          return;
-        }
-
-        tx.date = tz.TZDateTime.from(
-          tx.date.copyWith(
-            year: pickedDate.year,
-            month: pickedDate.month,
-            day: pickedDate.day,
-          ),
-          tx.date.location,
-        );
-      },
-    ),
-  );
-
-  Widget _buildTime(BuildContext context) => IntrinsicWidth(
-    child: TextFormField(
-      controller: tx.timeTC,
-      decoration: const InputDecoration(border: OutlineInputBorder()),
-      readOnly: true,
-      enabled: !saving,
-      onTap: () async {
-        final TimeOfDay? pickedTime = await showTimePicker(
-          context: context,
-          initialTime: tx.date.getTimeOfDay(),
-        );
-
-        if (pickedTime == null) {
-          return;
-        }
-
-        tx.date = tx.date.setTimeOfDay(pickedTime);
-      },
-    ),
-  );
 }
